removed adv logs for audio, moved types to types.js
