+++ rollplay/app/audio_management/types.js
@@ -1,0 +1,17 @@
+/*
+ * Audio Management Types
+ *
+ * Copyright (C) 2025 Matthew Davey
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ */
+
+export const PlaybackState = {
+  STOPPED: 'stopped',
+  PLAYING: 'playing',
+  PAUSED: 'paused'
+};
+
+export const ChannelType = {
+  BGM: 'bgm',
+  SFX: 'sfx'
+};
